Added event listener to show details upon mouseclick
loadDetails() should GET the Pokémon details using the URL from the Pokémon object in the parameter.
Once the GET request is complete, use then to return a JSON response.
then, assign some of the details you got from the response to the Pokémon in the pokemonList. Assign at least imgUrl and height.
Make sure both functions LoadList() and loadDetails() are assigned to keys with the same name in the returned object of your pokemonRepository.